pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Production RAG system for Indian legal statutes"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langgraph>=0.0.20",
    "sentence-transformers>=2.3.0",
    "faiss-cpu>=1.7.4",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rank-bm25>=0.2.2",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "google>=3.0.0",
    "openai>=2.15.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "google-generativeai>=0.8.6",
    "numpy>=1.26.4",
    "pypdf>=6.6.2",
]
